feat: add activity records to SslPair and split SslId into two VOs
The addition of the activity record in the CreateSslPair use case follows the same principles as previous cases. The need for a description is due to changes in the SslId VO, which is now split into two: SslPairId and SslCertificateId.

This change was necessary because the two constructors previously used the same VO but generated different contents, even though both were valid. This was an architectural error, considering that both values are indeed valid and should be used, but they belong to different contexts.